refactor: 💡 allow float vehicle volume

+++ packages/engine-ui/src/components/CreateVehicle.js
@@ -55,7 +55,7 @@ const CreateVehicle = ({ onSubmit }) => {
           : formState.timewindow,
       capacity: {
         weight: parseInt(formState.capacity.weight),
-        volume: parseInt(formState.capacity.volume),
+        volume: parseFloat(formState.capacity.volume),
       },
       startPosition: {
         ...formState.startPosition,
